video2numpy: separate FrameReader interface from read_vids function (since there will be more)
